segregated building workflow `VersionedDocs` from `__main__`
